rename: BridgeGame Class 패키지 위치 변경
PlayerSetting은 패키지 오류가 없고 BridgeGame의 오류가 있는지 확인하기 위해 변경
